+++ package.json
@@ -15,7 +15,7 @@
     "build": "vite build",
     "preview": "vite preview",
     "check": "svelte-check --tsconfig ./tsconfig.app.json && tsc -p tsconfig.node.json",
-    "postbuild": "cp index.dist.html dist/index.html"
+    "postbuild": "cp index.dist.html.txt dist/index.html"
   },
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^5.1.0",
