+++ src/lib/codeBlockHighlighter.ts
@@ -1,4 +1,7 @@
-import { codeToHtml } from 'shiki';
+import { createHighlighter } from 'shiki';
+import { createHighlighterCore } from '@shikijs/core';
+import { createOnigurumaEngine } from '@shikijs/engine-oniguruma'
+//import nord from "@shikijs/themes/dist/nord.mjs";
 
 /**
  * Highlights all <code class="language-xxx"> blocks by replacing them
@@ -9,6 +12,7 @@ import { codeToHtml } from 'shiki';
  * @param root - Optional root (ShadowRoot or HTMLElement). If not provided, autodetected.
  * @param theme - Shiki theme to use (default: 'catppuccin-latte')
  */
+
 export async function highlightAllCodeBlocks(
   root?: ShadowRoot | HTMLElement,
   theme: string = 'catppuccin-latte'
@@ -45,7 +49,7 @@ export async function highlightAllCodeBlocks(
     const rawCode = codeEl.textContent || '';
 
     try {
-      const highlightedHtml = await codeToHtml(rawCode, {
+      const highlightedHtml = await createShiki().codeToHtml(rawCode, {
         lang: language,
         theme: theme,
       });
@@ -64,4 +68,25 @@ export async function highlightAllCodeBlocks(
       console.error(`Error highlighting code block (lang: ${language}):`, error);
     }
   }
+}
+
+async function createShiki() {
+  const highlighter = await createHighlighterCore({
+    themes: [
+      // instead of strings, you need to pass the imported module
+      nord,
+      // or a dynamic import if you want to do chunk splitting
+      //import('@shikijs/themes/material-theme-ocean')
+    ],
+    langs: [
+      import('@shikijs/langs/javascript'),
+      // shiki will try to interop the module with the default export
+      () => import('@shikijs/langs/css'),
+      // or a getter that returns custom grammar
+      async () => JSON.parse(await fs.readFile('my-grammar.json', 'utf-8'))
+    ],
+    // `shiki/wasm` contains the wasm binary inlined as base64 string.
+    engine: createOnigurumaEngine(import('shiki/wasm'))
+  });
+  return highlighter;
 }
