+++ src/lib/components/ChoiceView.svelte
