+++ src/lib/components/Hint.svelte
