+++ src/lib/components/SequenceView.svelte
