+++ src/lib/components/SmoothResize.svelte
