--- src/registerIcons.ts
@@ -1,22 +1,0 @@
-import { library, config } from '@fortawesome/fontawesome-svg-core';
-import {
-    faArrowLeft,
-    faArrowRight,
-    faRedo,
-    faCheckDouble,
-    faCog,
-    faCircleNotch,
-} from '@fortawesome/free-solid-svg-icons';
-import { faLightbulb } from '@fortawesome/free-regular-svg-icons';
-
-export default function () {
-    config.autoAddCss = false;
-    library.add(
-        faArrowLeft,
-        faArrowRight,
-        faRedo,
-        faLightbulb,
-        faCheckDouble,
-        faCircleNotch
-    );
-}
